+++ package.json
@@ -33,6 +33,7 @@
         "i": "^0.3.7",
         "joi": "^17.13.3",
         "jsonwebtoken": "^9.0.2",
+        "net-snmp": "^3.12.0",
         "npm": "^10.2.4",
         "snmp-native": "^1.2.0"
     },
